moved the ros and gstreamer parts into individual classes in individual files
